pyproject.toml
[project]
name = "embedded-research-runtime"
version = "0.1.0"
description = "Lightweight research operations management platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "jinja2>=3.1",
    "typer>=0.12",
    "pydantic>=2.7",
    "python-frontmatter>=1.1",
    "markdown-it-py>=3.0",
]

[project.scripts]
err = "err.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/err"]

[tool.uv]
package = true
